refactor: Move coverage package
Move the coverage package from the public api into the internal impl package. These classes are not part of the API.
