+++ testing/camunda-process-test-java/src/test/java/io/camunda/process/test/api/ProcessEngineCoverageIT.java
@@ -19,10 +19,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.client.CamundaClient;
 import io.camunda.client.api.response.ProcessInstanceEvent;
-import io.camunda.process.test.api.coverage.core.CoverageCollector;
-import io.camunda.process.test.api.coverage.model.Coverage;
-import io.camunda.process.test.api.coverage.model.Run;
 import io.camunda.process.test.impl.assertions.CamundaDataSource;
+import io.camunda.process.test.impl.coverage.core.CoverageCollector;
+import io.camunda.process.test.impl.coverage.model.Coverage;
+import io.camunda.process.test.impl.coverage.model.Run;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
